refactor(forms): rename files related to metadata
Renames some files to reflect the property => metadata name change
