+++ packages/forms/signals/public_api.ts
@@ -15,7 +15,7 @@ export * from './src/api/async';
 export * from './src/api/control';
 export * from './src/api/field_directive';
 export * from './src/api/logic';
-export * from './src/api/property';
+export * from './src/api/metadata';
 export * from './src/api/structure';
 export * from './src/api/types';
 export * from './src/api/validation_errors';
