+++ packages/forms/signals/src/api/validators/min.ts
@@ -8,7 +8,7 @@
 
 import {computed} from '@angular/core';
 import {aggregateMetadata, metadata, validate} from '../logic';
-import {MIN} from '../property';
+import {MIN} from '../metadata';
 import {FieldPath, LogicFn, PathKind} from '../types';
 import {minError} from '../validation_errors';
 import {BaseValidatorConfig, getOption, isEmpty} from './util';
